Server --- set auth layer for all Note operations

+++ controllers/notes.js
@@ -5,8 +5,10 @@ module.exports = {
   // @desc      Get all notes
   // @access    Private
   index: async (req, res) => {
+    const userId = req.id;
+
     try {
-      const notes = await Note.find();
+      const notes = await Note.find({ user: userId }, "title body date");
       res.status(200).json(notes);
     } catch (err) {
       console.error("Error:", err.message);
@@ -17,9 +19,22 @@ module.exports = {
   // @desc      Create a note
   // @access    Private
   create: async (req, res) => {
+    const userId = req.id;
+    const { title, body } = req.body;
+
     try {
-      const note = await Note.create(req.body);
-      res.status(201).json(note);
+      const note = await Note.create({
+        user: userId,
+        title,
+        body
+      });
+
+      res.status(201).json({
+        _id: note._id,
+        date: note.date,
+        title: note.title,
+        body: note.body
+      });
     } catch (err) {
       console.error("Error:", err.message);
       res.status(500).json({ error: err.message });
@@ -29,10 +44,14 @@ module.exports = {
   // @desc      Get a note
   // @access    Private
   read: async (req, res) => {
+    const userId = req.id;
     const id = req.params.id;
 
     try {
-      const note = await Note.findById(id);
+      const note = await Note.findOne(
+        { _id: id, user: userId },
+        "date title body"
+      );
 
       res.status(200).json(note);
     } catch (err) {
@@ -44,11 +63,12 @@ module.exports = {
   // @desc      Edit a note
   // @access    Private
   update: async (req, res) => {
+    const userId = req.id;
     const id = req.params.id;
     const updatedNote = { ...req.body, date: Date.now() };
 
     try {
-      await Note.findByIdAndUpdate(id, updatedNote);
+      await Note.findOneAndUpdate({ _id: id, user: userId }, updatedNote);
 
       res.status(200).json({ msg: "Note updated successfully" });
     } catch (err) {
@@ -60,10 +80,11 @@ module.exports = {
   // @desc      Delete a note
   // @access    Private
   delete: async (req, res) => {
+    const userId = req.id;
     const id = req.params.id;
 
     try {
-      await Note.deleteOne({ _id: id });
+      await Note.deleteOne({ _id: id, user: userId });
 
       res.status(200).json({ msg: "Note deleted successfully" });
     } catch (err) {
